+++ lib/sidekiq/rails.rb
@@ -51,9 +51,8 @@ module Sidekiq
     end
 
     initializer "sidekiq.backtrace_cleaner" do
-      Sidekiq.configure_server do |config|
-        config[:backtrace_cleaner] = ->(backtrace) { ::Rails.backtrace_cleaner.clean(backtrace) }
-      end
+      Sidekiq.default_configuration[:backtrace_cleaner] =
+        ->(backtrace) { ::Rails.backtrace_cleaner.clean(backtrace) }
     end
 
     # This hook happens after all initializers are run, just before returning