+++ test/dummy/app/jobs/dummy_job.rb
@@ -1,0 +1,6 @@
+class DummyJob
+  include Sidekiq::Job
+
+  def perform
+  end
+end