+++ test/dummy/app/services/dummy_service.rb
@@ -1,0 +1,7 @@
+require_relative "../jobs/dummy_job"
+
+class DummyService
+  def self.do_something
+    DummyJob.perform_async
+  end
+end